+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 - Fix color related issues
 - Migrate DB to change COLUMN `hexColor:String` to `colorTag: EntryColorTag`
 - Extract all text to `strings.xml` for transliteration (except `autoTitle` in `JournalViewModel`)
+- Display entries in reverse order (latest first).
+- Updated move (drag) entry logic for reverse order in `JournalViewModel`.
+- MarkDownText support for lists, starting with `* ` and `+ `
+- MarkDownText support for inline links `<URL>` and `[URL Text](URL)`